Add structured logging support and suppress ASP.NET Core diagnostic logs
